Implemented Mazs feedback for pull requent
