--- demos/Lidar-demo/conda_win_env.txt
@@ -1,35 +1,0 @@
-name: D:\Projects\urc-intelligent-systems-2023\env
-channels:
-  - defaults
-dependencies:
-  - bzip2=1.0.8=he774522_0
-  - ca-certificates=2023.08.22=haa95532_0
-  - expat=2.5.0=hd77b12b_0
-  - libffi=3.4.4=hd77b12b_0
-  - openssl=3.0.11=h2bbff1b_2
-  - pip=23.3=py312haa95532_0
-  - python=3.12.0=h1d929f7_0
-  - setuptools=68.0.0=py312haa95532_0
-  - sqlite=3.41.2=h2bbff1b_0
-  - tk=8.6.12=h2bbff1b_0
-  - tzdata=2023c=h04d1e81_0
-  - vc=14.2=h21ff451_1
-  - vs2015_runtime=14.27.29016=h5e58377_2
-  - wheel=0.37.1=pyhd3eb1b0_0
-  - xz=5.4.2=h8cc25b3_0
-  - zlib=1.2.13=h8cc25b3_0
-  - pip:
-      - contourpy==1.1.1
-      - cycler==0.12.1
-      - fonttools==4.43.1
-      - kiwisolver==1.4.5
-      - matplotlib==3.8.0
-      - numpy==1.26.1
-      - packaging==23.2
-      - pillow==10.1.0
-      - pyparsing==3.1.1
-      - pyserial==3.5
-      - python-dateutil==2.8.2
-      - rplidar==0.9.2
-      - six==1.16.0
-prefix: D:\Projects\urc-intelligent-systems-2023\env
